# api.yml
openapi: 3.0.3
info:
  title: Sports Tournament Management API
  description: API for managing sports tournaments, teams, matches, and organization committees
  version: 1.0.0
servers:
  - url: 'https://api.example.com/v1'
    description: Production server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
      required:
        - message
        - code
paths:
...
  /stage/manage_groups:
    get:
      summary: Get group management page
      parameters:
        - name: tournamentId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Group management page
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                not_found:
                  value:
                    message: "Чемпионат не найден"
                    code: 404
                already_published:
                  value:
                    message: "Групповой этап уже опубликован"
                    code: 400
                no_access:
                  value:
                    message: "Нет доступа (пользователь – не старший организатор)"
                    code: 403
...